gpu-worker/chatterbox/requirements.txt
# Chatterbox TTS Worker Requirements
# Isolated container to avoid dependency conflicts

# =============================================================================
# Core ML Frameworks (versions required by chatterbox-tts)
# =============================================================================
torch==2.6.0
torchaudio==2.6.0

# =============================================================================
# Text-to-Speech
# =============================================================================
chatterbox-tts==0.1.6

# =============================================================================
# Self-hosted GPU & Storage
# =============================================================================
self-hosted==1.7.4
supabase==2.11.0
httpx==0.27.2

# =============================================================================
# Utilities
# =============================================================================
python-dotenv==1.0.1
scipy
rich==13.9.4
psutil==5.9.8
